Some new features.
Activation of the updateURL script when a module document is loaded.
Standardized way to encode a package (e.g.: "dbcsr/base") into something suitable as a file name prefix.
Printout of a JSON file for private but referenced (and so documented) symbols.
Printout of a JSON file with packages and module --> package mapping.